refactor : mutatian 기능을 커스텀 훅으로 분리
